requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-JWT-Extended==4.5.3
pyserial==3.5
waitress==2.1.2
Werkzeug==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.23
alembic==1.12.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
WTForms==3.1.1
PyJWT==2.8.0
cryptography==41.0.5
greenlet==3.0.1
Mako==1.3.0
typing-extensions==4.8.0
requests==2.31.0
pytz==2024.1
